refactor: dont need core or public. leave it just as src
