+++ engine/graphics/src/camera3d.cpp
@@ -1,7 +1,4 @@
 #include "graphics/camera3d.hpp"
-#include "graphics/window.hpp"
-#include "../core/constants.hpp"
-#include "utils/macros.hpp"
 #include <GLFW/glfw3.h>
 #include <glm/ext/matrix_clip_space.hpp>
 #include <glm/ext/matrix_transform.hpp>
