+++ engine/graphics/src/ring_buffer.hpp
