+++ engine/graphics/src/shapes.hpp
@@ -3,7 +3,6 @@
 
 namespace graphics {
 // clang-format off
-namespace internal {
 constexpr std::array<float, 18> SQUARE_VERTICES = {
     0, 0, 0,
     1, 0, 0,
@@ -13,6 +12,5 @@ constexpr std::array<float, 18> SQUARE_VERTICES = {
     0, 1, 0,
     0, 0, 0
     };
-}
 // clang-format on
 } // namespace graphics
